organize helper methods
